net/mlx5e: Do legacy SRIOV to flow offloads transitions under rtnl lock
vf_reps_work is removed and the transition is now done
from ethtool context under the rtnl lock.
{un}register_netdev is replaced with {un}register_netdevice
as we already hold the rtnl lock.
The free_netdev calls for the VF representors were moved to the
netdev destructors as we have to wait for unregister_netdevice to complete
before calling free_netdev and it completes only after rtnl_unlock.

Change-Id: Ic3f2866ac423a7573dd36794c547927a47a34c51
Signed-off-by: Ilya Lesokhin <ilyal@mellanox.com>