Revert "net/mlx4: Remove improper usage of dma_alloc_coherent()."
This reverts commit 1d03b6728d597525ca6f807e108ba3f423599995.
The reverted commit inserted a bug in the IB flow when creating qp, the
new way created the buffer as a list of pages, it caused a memory
corruption in the wqe errors and after that the device can get panic at
any place, the first indication for that is messages like:
"ib0: failed send event (status=2, wrid=90 vend_err 6f)", which happened
in the IPoIB driver in the completion for sends, (6f error indicates
big inline size, in the send flow there is no inline at all.)

Galia Magen found an easy way to reproduce: by running the gro_test.py
in one side with reloading the driver in the other side.

We decided to take out the commit because the commit itself is not the
way to fix the original issue, till we will have the decision it goes
out of the tree.

issue: 654545
issue: 668377
Change-Id: Id06833c06c6b87427cb453b4819b4d416e415553
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>