IB/core: Support accessing SA in virtualized environment
Per the ongoing standardisation process, when virtual HCAs are present
in a network, traffic is routed based on a destination GID. In order to
access the SA we use the well known SA GID.

This patch also adds a grh_required boolean field to ib_port_attr which
is used to report the verbs consumer whether this port is connected to a
virtualized network and needs to use GRH for communication. We clear the
port attributes struct before calling the hardware driver to make sure
the default remains that GRH is not required.

Issue: 702759
Change-Id: I474b46cd20a05210646a3c9db99b8e6f3b817250
Signed-off-by: Eli Cohen <eli@mellanox.com>