net/mlx5e: Pass input port representor to mlx5e_flow_act
Current code passes the PF when creating or destroying a flow, but then
looks up the source vport again. Since the source vport is always the
representor used by switchdev.

Just pass the representor we get from switchdev and use that as the
source vport for the flow.

Issue: none
Change-Id: Ib91db1d77607ff5f9ffd1ec1ccd6bb7538dc505e
Signed-off-by: Ilya Lesokhin <ilyal@mellanox.com>
Signed-off-by: Haggai Eran <haggaie@mellanox.com>