net/mlx5e: Set VST vlan insertion policy to overwrite
Currently we do insertion only if the VM packet didn't have vlan.

This is wrong, a VST VM may send a packet to a broadcast domain / tenant
it doesn't belong to by putting a different vlan and the we let it go.

Fix that by using insert or overwrite policy.

Issue: 681037
Change-Id: Ib6e2fcf89e00dde0cb9c5e8c4aceadf7f4c05955
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>