net/mlx5e: Split main flow table
Before this patch, the main flow table was used for two purposes:
1. Filter the packet dmac.
2. Classify the packet l3-l4 header in order to steer the packet to
the right RSS TIR.

This design is very complex, for each configured mac address we
have to add 11 rules (rule for each traffic type), the same if the
driver is inserted into promiscuous/allmulti mode.
In addition this scheme isn't scalability for future features like
aRFS or "foo over udp".

In order to simplify it, the main flow table is split to two flow
tables:
1. l2 table - filter the packet dmac address, if there is a match
we forward to the ttc flow table.

2. ttc table - classify the traffic type of the packet and steer the
packet to the right TIR.

In this new design, when new mac address is added, the driver added
just one flow rule(instead of 11).

Issue: 694100
Change-Id: I629956cfe1f4263d9ac2a82f2756c068bdd212fa
Signed-off-by: Maor Gottlieb <maorg@mellanox.com>