net/mlx5_en: Enabling aRFS mechanism
Accelerated RFS requiers that ntuple filtering is enabled via
ethtool.
When the ntuple filtering is enabled, we moidfy the l3_l4 ttc
rules to point on the aRFS flow tables and when the filtering
is disabled, we modify the l3_l4 ttc rules to point on the RSS
TIRs.
This patch sets the ndo_rx_flow_steer.

Issue: 693406
Change-Id: Ibb78c6f2b8de9dbf50464e806c68170765f3eca2
Signed-off-by: Maor Gottlieb <maorg@mellanox.com>