net/mlx5: Use reserved_at_offset in mlx5_ifc.h
mlx5_ifc.h is a header file representing the API and ABI between
the driver and the Mellanox ConnectX4 and Connect-IB firmware and
hardware. This file is used from both the mlx5_ib and mlx5_core
drivers.

Previously, this file used incrementing counter to indicate reserved
fields, for example:

struct mlx5_ifc_odp_per_transport_service_cap_bits {
        u8         send[0x1];
        u8         receive[0x1];
        u8         write[0x1];
        u8         read[0x1];
        u8         reserved_0[0x1];
        u8         srq_receive[0x1];
        u8         reserved_1[0x1a];
};

If developer A from the net tree implements a feature that uses
reserved0, he replaces it with featureA and renames reserved_1 to
reserved_0. In the same kernel cycle, developer B could implement a
feature that uses reserved_1 and split it to featureB and a smaller
reserved1 field. This will cause a conflict.

The source of this conflict is that developerA changed *all* reserved
fields after his feature.

As Linus suggested, this patch changes this struct to:

struct mlx5_ifc_odp_per_transport_service_cap_bits {
	u8         send[0x1];
	u8         receive[0x1];
	u8         write[0x1];
	u8         read[0x1];
	u8         reserved_at_4[0x1];
	u8         srq_receive[0x1];
	u8         reserved_at_6[0x1a];
};

This makes the conflicts more rare and the preserves the locality of
the changes.

Change-Id: Icb778d70c4322f06233304fc9e01a34bdda10232
Signed-off-by: Matan Barak <matanb@mellanox.com>