IB/core: Fix input len in vendor specific part of create_qp/srq cmd
Currently, the inlen field of the vendor's part of the command
doesn't match the in part. This happens because the inlen
accommodates ib_uverbs_cmd_hdr which is deducted from the in buffer.
Fixing this by aligning the inlen to the correct length.
All vendor drivers either assumed that inlen >= sizeof(vendor_uhw_cmd)
or just failed wrongly (mlx5).

issue: 697949
Change-Id: I5453c80491f99cc40f02af8e7d1d5a409b186be0
Signed-off-by: Majd Dibbiny <majd@mellanox.com>