net/mlx5: Introduce device queue counters
A queue counters can collect several statisics for one or more
hardware queues (QPs, SQs, etc ..) that the counter is attached to.

Here we add device commands to alloc/query/dealloc and manage queue
counters, the device has a limited count of this resource and we devided
the qouta between IB/ETH clients in a way that gurantees both will get
thier share of the pie.

For ethernet we preserve upto 64 sets of queue counters, IB will take
the rest.

Issue: none
Change-Id: I06dd63c4b793fc8375aceb6ecda165d6e371b254
Signed-off-by: Rana Shahout <ranas@mellanox.com>
Signed-off-by: Tariq Toukan <tariqt@mellanox.com>