net/mlx5: Introduce device queue counters
A queue counter can collect several statistics for one or more
hardware queues (QPs, RQs, etc ..) that the counter is attached to.

For Ethernet it will provide an "out of buffer" counter which collects
the number of all packets that are dropped due to lack of software buffers.

Here we add device commands to alloc/query/dealloc and manage queue
counters, the device has a limited count of this resource and we divided
the quota between IB/ETH clients in a way that guarantees both will get
their share of the pie.

For ethernet we preserve up to 64 sets of queue counters, IB will take
the rest.

Issue: 693567
Change-Id: Ie83201010cc04fd67c39ee893bb47fddeedde476
Signed-off-by: Tariq Toukan <tariqt@mellanox.com>
Signed-off-by: Rana Shahout <ranas@mellanox.com>