net/mlx5e: Consider IRQ affinity changes in NAPI poll
Under heavy network traffic load we may be in NAPI poll mode for
a long time thus IRQ affinity changes do not affect NAPI as desired.

So once in a NAPI poll session we check the channel IRQ affinity and
if it points to a different CPU core than the current we exit polling
and the following device interrupt will fix the NAPI affinity.

Change-Id: I9a5584d36ff1321e90cf3509f81ad0e1f2caa430
Signed-off-by: Achiad Shochat <achiad@mellanox.com>
Signed-off-by: Tariq Toukan <tariqt@mellanox.com>