net/mlx5e: Implement CQE compression feature
CQE compression feature is meant to save PCIe bandwidth by
compressing few CQEs into smaller amount of bytes on PCIe.
CQE compression is selectively enabled per CQ.
It is not meant to be used by Latency Highly-sensitive applications,
as CQEs may be delayed for compressing.

CQE compression terms:
CQE slot - location in memory of a single CQE in CQ buffer.
CQE compression block - this is the section in CQ buffer where the compression occurs.
CQE compression title - this object is put is the first CQE-slot of a CQE compression block.
It includes a field that indicates it is a CQE compression title.
Mini CQE - a 8-byte object that holds some data needed to extract a single compressed CQE.
Mini CQE array - a collection of 8 Mini CQEs,
laid out on a single CQE-slot in the CQE compression block.

Decompressing CQEs:
First CQE slot holds the CQE compression title.
The successor CQE slot holds the first Mini CQE array.
In case compression occurred for more than 8 CQEs,
additional Mini CQE arrays will be located in strides of 8 CQE slots
starting from the CQE compression title.
When SW polls a CQ and identifies a CQE compression title,
it should take the contents of the title,
and copy it with little modifications into the CQ buffer,
to its own CQE slot and to the CQE slots following the CQE compression title.
CQE compression title includes the number of CQE slots it should be spread out to.
In each CQE slot where CQE compression title is copied to,
a respective Mini CQE should be used, in order to modify some of the copied contents.
Locations in the CQE slot for Mini CQE dwords are determined according to CQ configuration.

issue: 567013

Change-Id: I733515cd51a5baea7c0c3bf1b80912bac81a3e3a
Signed-off-by: Eugenia Emantayev <eugenia@mellanox.com>
Signed-off-by: Tariq Toukan <tariqt@mellanox.com>