net/mlx5e: Add basic encap support
The code maintains mlx5e_encap_entry for each encap header
it has encountered. This entry maintains a linked list
of all the flows sharing the encap header, when the last
flow is removed from the list the entry is removed.

The actual encap_header is allocated in the hardware only
if we have layer 2 info when the mlx5e_encap_entry was created.
The field valid_encap tells us whether the header is in the hardware
and encap_id tells us the encap_id that was allocated.
While the flow is in the driver, the driver holds
a reference on the neighbour.

When a new flow with encap is inserted the code first check
if the required header exist, if it does the encap
is shared, otherwise a new mlx5e_encap_entry is created.
The flow is inserted to the FDB only if the corresponding encap
is valid. if the encap is invalid we don't insert the flow
to the FDB but keep it in the driver with
flow_index == INVALID_FLOW_IX.

The next patch will add code to add and remove flows and encap headers
in response to neighbour updates.

Issue: None
Change-Id: Icfeac91442923d6fc675ce3f8dd22cde32461d4a
Signed-off-by: Ilya Lesokhin <ilyal@mellanox.com>