net/mlx4_core: Allow resetting vf admin mac to zero
The VF administrative mac addresses (stored in the PF driver) are
initialized to zero when the PF driver starts up.
These addresses may be modified in the PF driver through ndo calls
initiated by ip tool or libvirt.

When the VF is restarted (or initially started) the administrative
mac address stored in the PF for that VF is sent to the VF as its
mac address.
If the sent mac is zero, the VF generates a random mac for itself.
If the mac is a valid unicast mac, that mac is used.

While we allow the PF/host to change the VF admin mac address from
zero to a valid unicast mac, we do not allow restoring the zero VF
admin mac if we change our minds later. We currently only allow changing
this mac to a different unicast mac. This leads to problems when
libvirt scripts are used to deal with VF mac addresses.

Fix this by allowing resetting a VF administrative MAC back to zero.

issue: 736559
Change-Id: Ie988a92a9cccdb18f9a06eed8ea14cec75bad8c4
Fixes: 8f7ba3ca12f6 ("net/mlx4: Add set VF mac address support")
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>