net/mlx5e: Provide correct packet/bytes statistics
[Commit faf4478bf8b7b5e6c94cb07a52a00e4d1cd647c6 upstream]

Using the HW VPort counters for traffic (rx/tx packets/bytes)
statistics is wrong. This is because frames dropped due to steering or
out of buffer will be counted as received. To fix that, we move to use
the packet/bytes accounting done by the driver for what the netdev
reports out.

Issue: 735895
Change-Id: Ib516d1f592b0495019d1523e449251cccc468501
Fixes: f62b8bb8f2d3 ('net/mlx5: Extend mlx5_core to support [...]')
Signed-off-by: Gal Pressman <galp@mellanox.com>