net/mlx5e: Register netevent notifier
Register a netevent notifier and use mlx5e_pf_switchdev_ops
to identify events that are relevant to the PF.
When we receive a neighbour update, we check
if e->valid_encap != neighbour_valid(n) or
if the mac address of the neighbour changed.
If thats the case we add the relevant encap header entry
to update_encaps_list and wake up a thread to process
this list under the rtnl lock.
For each encap entry on the list we remove all the flows
and the encap header from the hardware
and then try to recreate the header and reinsert the flows.
Additionally, if a neighbour with offloaded flows becomes
stale we use neigh_event_send() to verify it still valid.
Note that this verification happens even if there if there
is no traffic on those flows.

Issue: None
Change-Id: I36db2c199c4b927455db68907a1d4509e6b0e2c1
Signed-off-by: Ilya Lesokhin <ilyal@mellanox.com>