IB/mlx5: Added support for re-registration of MRs
This patch adds support for re-registration of memory regions in MLX5.
The functionality is basically the same as deregister followed by
register, but attempts to reuse the existing resources as much as
possible.
Original memory keys are kept if possible, saving the need to
communicate new ones to remote peers.

issue: 681015

Change-Id: Id88473c45324affaa0d86e6cbe0f64d130d787fd
Signed-off-by: Noa Osherovich <noaos@mellanox.com>