net/mlx5e: Support RX multi-packet WQE
Introduce the feature of multi-packet RX Work Queue Element
(multi-packet WQE, MPWQE, referred also as Striding RQ),
in which WQEs are larger and serve multiple packets each.
Every WQE consists of strides of the same size, every received
packet is aligned to a beginning of a stride and is written to
consecutive strides within a WQE.
In the regular approach, each regular WQE is big enough to be capable of serving
one received packet of any size, making it very wasteful when dealing with small packets.
For its flexibility, MPWQE allows a better memory utilization (implying improvements in CPU util and packet rate) as
packets consume strides according to their size, preserving the rest of the WQE available for other
packets.

Performance numbers: /* TODO */

Issue: 677953
Issue: 704476
Change-Id: I942fb462754cb4234f1caba55c3f4baf5a62df7e
Signed-off-by: Tariq Toukan <tariqt@mellanox.com>
Signed-off-by: Achiad Shochat <achiad@mellanox.com>