net/mlx5e: The port MTU shouldn't be used as a lower bound for the netdev MTU
The port MTU should be used only as an upper bound for the netdev MTU.
In transmit there is no point in using the port MTU as a lower bound.
In receive the driver is allowed to drop, truncate or pass oversized
packets. Since the Hardware has no issue dropping packets that
exceed the receive buffer size, again there is no reason to limit the
netdev MTU.

Issue: 650981
Change-Id: Id01f74a9fa38d85a2e27ccc99ed43dea2957f2da
Signed-off-by: Ilya Lesokhin <ilyal@mellanox.com>