IB/core: Add subnet prefix to port info
The subnet prefix is a part of the port_info MAD returned and should be
available at the ib_port_attr struct. We define it here and implement
for all the hardware drivers. The subnet prefix is required when
creating the address vector to access the SA in networks where GRH must
be used.

Issue: 702759
Change-Id: I07ef7f7e0596a507793c0cfaf97b1c8e3fa32cc0
Signed-off-by: Eli Cohen <eli@mellanox.com>