net/mlx5: E-Switch, Replace vport spin lock with synchronize_irq()
Vport spin lock can be replaced with synchronize_irq() in the right
place, this will remove the need of locking inside irq context.
Locking in esw_enable_vport is not required since vport events are yet
to be enabled, and at esw_disable_vport it is sufficient to
synchronize_irq() to guarantee no further vport events handlers will be
scheduled.

Issue: 747240
Change-Id: I7183f4f04b586fe8e4f4dc26fdd605a325bba746
Signed-off-by: Mohamad Haj Yahia <mohamad@mellanox.com>