net/mlx4_core: Set EQ affinity hint to local NUMA CPUs
Use CPUs on the close NUMA first when setting the EQ affinity hints.
Previously CPUs would be used in order, without regard to the NUMA node.
If the device was on a node with high numbered CPUs and there were
fewer EQs than CPUs then all EQ could be hinted to off node CPUs.

Change-Id: I2daa634c8de9703d3fd1d44dbcf2d451c4ff1ddc
Signed-off-by: Daniel Jurgens <danielj@mellanox.com>
Signed-off-by: Tariq Toukan <tariqt@mellanox.com>
Signed-off-by: Saeed Mahameed <saeedm@mellanox.com>