IB/mlx5: Support RAW Ethernet when RoCE is disabled
On some environments, such as certain SRIOV VF configurations, RoCE is
not supported for an Ethernet port. Currently, the driver will not
open IB device on that port.

This is problematic, since we do want user-space RAW Ethernet
(RAW_PACKET QPs) functionality to remain in place. For that end,
enhance the relevant driver flows such that we do open an IB device
in that case but don't advertize RoCE support towards the IB stack.

issue: 754137
Change-Id: Id35fd837fa644d2c9bc37a76613ad151a8fc17c8
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>