net/mlx5e: Put send-to-vport FDB steering rules after offloaded rules
Put the steering group used for send-to-vport rules to be after the
groups used for offloaded rules. This is possible since sent-to-vport
rules always match on source vport=0 which only holds for packets sent
from OVS over the PF or the reps.

This should reduce at least one HW steering FDB hop for the sriov
offloads mode.

issue: none
Change-Id: I7d00237f5775c21e08ad05a6c0161313461f3807
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>