mlx5e: Use random MAC address for representors
Having the same MAC address for VFs and their representors causes
problems with existing udev rules that only look at the mac address will
match both the VF and the representor. This results in udev
indeterministically assigning the desired netdev name to one of them. In
addition, using DHCP accidentally on the representor would result in the
DHCP server assigning it the VF's address.

Today the veth and tun/tap drivers assign a random mac to their netdevs.
libvirtd assigns a matching MAC to a VM interface and its tap device by
changing the first byte to 0xfe (see explanation [1]).

Randomize the representor MACs and leave it for a future patch of 
libvirt (adding representor support) to change the MAC of the 
representor in user-space.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=571991

Change-Id: I159b6acedc45ae59e0a508e2f6bef229ab736114
Signed-off-by: Haggai Eran <haggaie@mellanox.com>