net/mlx5e: Avoid offloading ARPs
ARP/RARP packets are OTOH used for protocol slow-path and OTOH capture
a flow group in the offloads logic, which means that offloading them
would potentially add more hops to the steering of non ARP packets,
and reduce performance. Avoid offloading them so we don't get perf drops
as of that.

Issue: none
Change-Id: I55103b86e9b41ef6aaf5fcaffd617eee91541d24
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: Ilya Lesokhin <ilyal@mellanox.com>