{net, IB}/mlx5: Refactor internal SRQ API
Currently, the SRQ API uses the obsolete mlx5_*_srq_mbox_{in,out}
structs which limit the ability to pass the SRQ attributes between
net and IB parts of the driver.

This patch changes the SRQ API so as to use auto-generated structs
and provides a better way to pass attributes which will be in use by
coming features.

Issue: 767293
Change-Id: I28997d61c93e2b31f57659cfd2107392079703a2
Signed-off-by: Artemy Kovalyov <artemyko@mellanox.com>