net/mlx5e: Allow offloading no fragment match when working in L2 mode
When OVS works in L2 forwarding mode (normal action)
it adds a no fragment match criteria.
The reason for the match is unclear,
and we don't want to penalize all the users that
work in MLX5_INLINE_MODE_L2, so we ignore it.

The side effect of this is that the user can't treat fragmented
and non fragmented traffic differently. If we a vm is allowed
to send non fragmented traffic then it will also be able to
send non fragmented traffic.

Ideally we would remove the no fragment match entirely, but
then we might get 3 different rules with the same match and
the driver doesn't support this at this point.

Issue: None
Change-Id: I755994af06a057ba26323427dbfb604f2efb80dd
Signed-off-by: Ilya Lesokhin <ilyal@mellanox.com>